change type of cost and gram(double to BigDecimal)

+++ src/online/bakery/CondimentDecorator.java
@@ -1,14 +1,16 @@
 package online.bakery;
 
+import java.math.BigDecimal;
+
 public abstract class CondimentDecorator extends Sweets {
 
-    public void setGrams(double grams) {
+    public void setGrams(BigDecimal grams) {
         this.Grams = grams;
     }
 
     public abstract String getDescription();
 
-    public double getGrams() {
+    public BigDecimal getGrams() {
         return Grams;
     }
